refactor(copilot_task_submit/transcript): merge inline tests into tests_transcript.rs (#1266)
Inline test module in transcript.rs is moved into existing tests_transcript.rs
as a nested mod transcript_inline {} submodule.

- src/copilot_task_submit/transcript.rs: 495 → 356 LOC (under 400-line rule)
- src/copilot_task_submit/tests_transcript.rs: 212 → 354 LOC
- Widened to pub(super): is_ignorable_copilot_ui_fragment

Pure refactor: zero logic changes, all 17 tests pass.

Closes 14/70 of #1266 (this loop iteration), 23 PRs cumulative since #1285.
