+++ src/copilot_task_submit/transcript.rs
@@ -308,7 +308,7 @@ fn canonicalize_visible_fragment(fragment: &str, flow: &CopilotSubmitFlowAsset) 
     Some(fragment.to_string())
 }
 
-fn is_ignorable_copilot_ui_fragment(fragment: &str) -> bool {
+pub(super) fn is_ignorable_copilot_ui_fragment(fragment: &str) -> bool {
     fragment.eq("GitHub Copilot")
         || fragment.starts_with("GitHub Copilot v")
         || fragment.starts_with("itHub Copilot")
@@ -352,144 +352,4 @@ pub(super) fn copilot_transcript_preview(
     }
 
     normalized
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::copilot_task_submit::types::{
-        CopilotSubmitFlowAsset, PositiveCheckpointKind, TranscriptCheckpointScan,
-    };
-    use std::path::PathBuf;
-
-    fn test_flow() -> CopilotSubmitFlowAsset {
-        CopilotSubmitFlowAsset {
-            launch_command: "copilot-cli".into(),
-            working_directory: PathBuf::from("."),
-            wait_timeout_seconds: 30,
-            startup_banner: "Welcome to Copilot".into(),
-            guidance_checkpoint: "Ready for input".into(),
-            submit_hint: "Press Enter to submit".into(),
-            post_submit_checkpoint: Some("Task completed".into()),
-            trust_prompt: None,
-            wrapper_error_signal: None,
-            workflow_noise_signals: vec![],
-            payload_id: "test-payload".into(),
-            payload: "do something".into(),
-        }
-    }
-
-    #[test]
-    fn classify_startup_ready_on_ordered_sequence() {
-        let mut scan = TranscriptCheckpointScan::default();
-        scan.record_checkpoint(PositiveCheckpointKind::StartupBanner, "banner", 0);
-        scan.record_checkpoint(PositiveCheckpointKind::Guidance, "guidance", 1);
-        let status = classify_startup(&scan, false);
-        assert_eq!(status, StartupStatus::Ready);
-    }
-
-    #[test]
-    fn classify_startup_wait_on_empty_scan() {
-        let scan = TranscriptCheckpointScan::default();
-        let status = classify_startup(&scan, false);
-        assert_eq!(status, StartupStatus::Wait);
-    }
-
-    #[test]
-    fn classify_startup_unsupported_on_wrapper_error() {
-        let scan = TranscriptCheckpointScan {
-            has_wrapper_error: true,
-            ..Default::default()
-        };
-        let status = classify_startup(&scan, false);
-        assert_eq!(status, StartupStatus::Unsupported("copilot-wrapper-error"));
-    }
-
-    #[test]
-    fn classify_startup_unsupported_on_trust_prompt() {
-        let scan = TranscriptCheckpointScan {
-            has_trust_prompt: true,
-            ..Default::default()
-        };
-        let status = classify_startup(&scan, false);
-        assert_eq!(
-            status,
-            StartupStatus::Unsupported("trust-confirmation-required")
-        );
-    }
-
-    #[test]
-    fn classify_startup_exited_with_sequence_gives_exited_early() {
-        let mut scan = TranscriptCheckpointScan::default();
-        scan.record_checkpoint(PositiveCheckpointKind::StartupBanner, "banner", 0);
-        scan.record_checkpoint(PositiveCheckpointKind::Guidance, "guidance", 1);
-        let status = classify_startup(&scan, true);
-        assert_eq!(status, StartupStatus::Unsupported("process-exited-early"));
-    }
-
-    #[test]
-    fn classify_submit_success_on_post_submit_checkpoint() {
-        let mut scan = TranscriptCheckpointScan::default();
-        scan.record_checkpoint(PositiveCheckpointKind::PostSubmitCheckpoint, "done", 5);
-        let status = classify_submit(&scan, false);
-        assert_eq!(status, SubmitStatus::Success);
-    }
-
-    #[test]
-    fn classify_submit_wait_on_empty() {
-        let scan = TranscriptCheckpointScan::default();
-        let status = classify_submit(&scan, false);
-        assert_eq!(status, SubmitStatus::Wait);
-    }
-
-    #[test]
-    fn classify_submit_timeout_returns_copilot_wrapper_error() {
-        let scan = TranscriptCheckpointScan {
-            has_wrapper_error: true,
-            ..Default::default()
-        };
-        let flow = test_flow();
-        assert_eq!(
-            classify_submit_timeout(&scan, &flow),
-            "copilot-wrapper-error"
-        );
-    }
-
-    #[test]
-    fn scan_transcript_lines_detects_banner_and_guidance() {
-        let flow = test_flow();
-        let lines = vec!["Welcome to Copilot", "Ready for input"];
-        let scan = scan_transcript_lines(lines, &flow);
-        assert!(scan.has_banner());
-        assert!(scan.has_guidance());
-        assert!(scan.has_ordered_startup_sequence());
-    }
-
-    #[test]
-    fn scan_transcript_lines_detects_post_submit() {
-        let flow = test_flow();
-        let lines = vec!["Task completed"];
-        let scan = scan_transcript_lines(lines, &flow);
-        assert!(scan.has_post_submit_checkpoint());
-    }
-
-    #[test]
-    fn is_ignorable_copilot_ui_fragment_github_copilot() {
-        assert!(is_ignorable_copilot_ui_fragment("GitHub Copilot"));
-        assert!(is_ignorable_copilot_ui_fragment("GitHub Copilot v1.2.3"));
-        assert!(is_ignorable_copilot_ui_fragment("Tip: do something"));
-    }
-
-    #[test]
-    fn is_ignorable_copilot_ui_fragment_normal_text() {
-        assert!(!is_ignorable_copilot_ui_fragment("Hello world"));
-    }
-
-    #[test]
-    fn copilot_transcript_preview_truncates() {
-        let flow = test_flow();
-        let fragments: Vec<String> = (0..200).map(|i| format!("line-{i}")).collect();
-        let preview = copilot_transcript_preview(&fragments, &flow);
-        assert!(preview.len() <= 515); // 512 + "..."
-    }
 }
